Time: 0 ms (100.00%) | Memory: 8.4 MB (36.28%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -19,7 +19,7 @@ public:
         }
     }
     double myPow(double x, int n) {
-            double num = n;
+            long num = n;
             if(n< 0){
                 return (1.0/power(x,-num));
             }
